+++ src/Plugin/CKEditorPlugin/TemplateWidget.php
@@ -8,6 +8,7 @@ namespace Drupal\ckeditor_template_widget\Plugin\CKEditorPlugin;
 
 
 use Drupal\ckeditor\CKEditorPluginContextualInterface;
+use Drupal\ckeditor\CKEditorPluginCssInterface;
 use Drupal\ckeditor\CKEditorPluginInterface;
 use Drupal\Component\Plugin\PluginBase;
 use Drupal\editor\Entity\Editor;
@@ -20,7 +21,7 @@ use Drupal\editor\Entity\Editor;
  *   label = @Translation("Templates widget")
  * )
  */
-class TemplateWidget extends PluginBase implements CKEditorPluginInterface, CKEditorPluginContextualInterface {
+class TemplateWidget extends PluginBase implements CKEditorPluginInterface, CKEditorPluginContextualInterface, CKEditorPluginCssInterface {
 
   /**
    * Checks if this plugin should be enabled based on the editor configuration.
@@ -116,5 +117,25 @@ class TemplateWidget extends PluginBase implements CKEditorPluginInterface, CKEd
   public function getConfig(Editor $editor) {
 
     return array();
+  }
+
+  /**
+   * Retrieves enabled plugins' iframe instance CSS files.
+   *
+   * Note: this does not use a Drupal asset library because this CSS will be
+   * loaded by CKEditor, not by Drupal.
+   *
+   * @param \Drupal\editor\Entity\Editor $editor
+   *   A configured text editor object.
+   *
+   * @return string[]
+   *   An array of CSS files. This is a flat list of file paths relative to
+   *   the Drupal root.
+   */
+  public function getCssFiles(Editor $editor) {
+
+    return array(
+      drupal_get_path('module', 'system') . '/css/components/clearfix.module.css'
+    );
   }
 }