+++ CHANGELOG.md
@@ -6,3 +6,6 @@ All notable changes to this project will be documented in this file.
 
 - CKEditor Template widget plugin. Drupal CKEditor plugin. Convert multiple
 columns templates into widget.
+
+- Use BEM for classes. Added require property to CKEditor plugin. Load 
+clearfix.module.css into CKEditor. Added full-page slide template.